Move parallel_chisqr to chisqr utitlites.
